Initial conversion to Backdrop
This handles the basic conversion to Backdrop, but the differences in how
the menu and block systems work means that this is still not functional.
The _block_rehash() function, for instance, has no equivalent so there may
be large chunks of code that need swapped out to work with Backdrop config.